+++ apps/meteor/app/livechat/server/lib/utils.ts
@@ -1,0 +1,5 @@
+import { RoutingManager } from './RoutingManager';
+
+export function showConnecting() {
+	return RoutingManager.getConfig()?.showConnecting || false;
+}
