+++ apps/meteor/server/features/EmailInbox/EmailInbox_Incoming.ts
@@ -17,6 +17,7 @@ import { notifyOnMessageChange } from '../../../app/lib/server/lib/notifyListene
 import { Livechat as LivechatTyped } from '../../../app/livechat/server/lib/LivechatTyped';
 import { QueueManager } from '../../../app/livechat/server/lib/QueueManager';
 import { setDepartmentForGuest } from '../../../app/livechat/server/lib/departmentsLib';
+import { sendMessage } from '../../../app/livechat/server/lib/messages';
 import { settings } from '../../../app/settings/server';
 import { i18n } from '../../lib/i18n';
 
@@ -146,7 +147,7 @@ export async function onEmailReceived(email: ParsedMail, inbox: string, departme
 	const rid = room?._id ?? Random.id();
 	const msgId = Random.id();
 
-	LivechatTyped.sendMessage({
+	sendMessage({
 		guest,
 		message: {
 			_id: msgId,
