+++ packages/apps-engine/src/server/ProxiedApp.ts
@@ -73,7 +73,7 @@ export class ProxiedApp {
             // Range of JSON-RPC error codes: https://www.jsonrpc.org/specification#error_object
             if (e.code >= -32999 || e.code <= -32000) {
                 // we really need to receive a logger from rocket.chat
-                console.error('JSON-RPC error received: ', e);
+                console.error('JSON-RPC error received: ', JSON.stringify(e));
             }
         }
     }
